6-size.c

+++ 0x00-hello_world/6-size.c
@@ -12,11 +12,11 @@ int main(void)
 	long long int lli;
 	float f;
 
-	printf("size of a char: %c bytes\n", sizeof(c));
-	printf("size of an int: %i bytes\n", sizeof(i));
-	printf("size of a long int: %li bytes\n", sizeof(li));
-	printf("size of a long long int: %lli bytes\n", sizeof(lli));
-	printf("size of a float: %f bytes\n", sizeof(f));
+	printf("size of a char: %c bytes\n", (unsigned long)sizeof(c));
+	printf("size of an int: %i bytes\n", (unsigned long)sizeof(i));
+	printf("size of a long int: %li bytes\n", (unsigned long)sizeof(li));
+	printf("size of a long long int: %lli bytes\n", (unsigned long)sizeof(lli));
+	printf("size of a float: %f bytes\n", (unsigned long)sizeof(f));
 
 	return (0);
 }
